457 bump bevy version to 0.15.3 and rust edition to 2024

+++ cosmos_client/src/structure/systems/laser_cannon_system.rs
@@ -38,7 +38,7 @@ fn apply_shooting_sound(
         let location = *ship_location;
         let translation = ship_global_transform.translation();
 
-        let idx = rand::random::<usize>() % audio_handles.0.len();
+        let idx = rand::random::<u64>() as usize % audio_handles.0.len();
 
         let handle = audio_handles.0[idx].clone_weak();
